# openapi.yaml
openapi: 3.0.3
info:
  title: Coleccion Musical Backend API
  version: '1.0.0'
  description: |
    REST API for managing musical artists and albums stored in PostgreSQL.
    All responses share a common envelope with `status`, `message`, optional `data`, and `timestamp` fields.
servers:
  - url: http://localhost:3000
    description: Local development server
components:
  schemas:
    ApiResponse:
      type: object
      properties:
        status:
          type: string
          example: success
        message:
          type: string
          example: Resource retrieved
        data:
          type: object
          nullable: true
        timestamp:
          type: string
          format: date-time
          example: 2025-09-29T21:30:00.000Z
    Artist:
      type: object
      properties:
        id_artista:
          type: integer
          example: 26
        nombre:
          type: string
          example: Ed Maverick
        genero_musica:
          type: string
          example: Independiente
      required:
        - id_artista
        - nombre
        - genero_musica
    Album:
      type: object
      properties:
        id_album:
          type: integer
          example: 12
        titulo_album:
          type: string
          example: Eduardo
        anio_album:
          type: integer
          example: 2021
        id_artista:
          type: integer
          example: 26
      required:
        - id_album
        - titulo_album
        - anio_album
        - id_artista
    ArtistCreateInput:
      type: object
      required:
        - nombre
        - genero_musica
      properties:
        nombre:
          type: string
          example: Ed Maverick
        genero_musica:
          type: string
          example: Independiente
    ArtistUpdateInput:
      type: object
      properties:
        nombre:
          type: string
        genero_musica:
          type: string
    AlbumCreateInput:
      type: object
      required:
        - titulo_album
        - anio_album
        - id_artista
      properties:
        titulo_album:
          type: string
          example: Eduardo
        anio_album:
          type: integer
          example: 2021
        id_artista:
          type: integer
          example: 26
    AlbumUpdateInput:
      type: object
      properties:
        titulo_album:
          type: string
        anio_album:
          type: integer
        id_artista:
          type: integer
paths:
...
  /api/artistas:
    get:
      summary: List artists
      responses:
        '200':
          description: Artists retrieved
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - properties:
                      data:
                        type: object
                        properties:
                          artistas:
                            type: array
                            items:
                              $ref: '#/components/schemas/Artist'
              examples:
                sample:
                  summary: Artists example
                  value:
                    status: success
                    message: Artists retrieved
                    data:
                      artistas:
                        - id_artista: 3
                          nombre: Adele
                          genero_musica: Soul
                        - id_artista: 18
                          nombre: Ariana Grande
                          genero_musica: Pop
                        - id_artista: 5
                          nombre: Billie Eilish
                          genero_musica: Electropop
                    timestamp: '2025-09-29T05:07:55.338Z'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
  /api/agregar_artista:
    post:
      summary: Create artist
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ArtistCreateInput'
      responses:
        '201':
          description: Artist created
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - properties:
                      data:
                        type: object
                        properties:
                          artista:
                            $ref: '#/components/schemas/Artist'
        '400':
          description: Missing fields
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '409':
          description: Duplicate artist name
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '500':
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
  /api/artistas/{id}:
    patch:
      summary: Update artist
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ArtistUpdateInput'
      responses:
        '200':
          description: Artist updated
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - properties:
                      data:
                        type: object
                        properties:
                          artista:
                            $ref: '#/components/schemas/Artist'
        '400':
          description: Invalid payload
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '404':
          description: Artist not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '409':
          description: Duplicate artist name
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '500':
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'


    delete:
      summary: Delete artist
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Artist deleted
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - properties:
                      data:
                        type: object
                        properties:
                          artista:
                            $ref: '#/components/schemas/Artist'
              examples:
                artistDeleted:
                  summary: Eliminacion exitosa
                  value:
                    status: success
                    message: Artista eliminado
                    data:
                      artista:
                        id_artista: 26
                        nombre: Ed Maverick
                        genero_musica: Independiente
                    timestamp: '2025-09-30T12:00:00Z'
        '400':
          description: Invalid id
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
              examples:
                invalidId:
                  summary: id no valido
                  value:
                    status: error
                    message: id debe ser un numero entero
                    data:
                      field: id
                    timestamp: '2025-09-30T12:00:00Z'
        '404':
          description: Artist not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
              examples:
                notFound:
                  summary: Artista inexistente
                  value:
                    status: error
                    message: Artista no encontrado
                    data:
                      id_artista: 99
                    timestamp: '2025-09-30T12:00:00Z'
        '409':
          description: Artist has related records
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
              examples:
                hasDependencies:
                  summary: El artista tiene dependencias
                  value:
                    status: error
                    message: No se puede eliminar el artista porque tiene registros asociados
                    data:
                      id_artista: 26
                    timestamp: '2025-09-30T12:00:00Z'
        '500':
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
              examples:
                serverError:
                  summary: Error inesperado
                  value:
                    status: error
                    message: Error al eliminar artista
                    data:
                      error: Descripcion del error
                    timestamp: '2025-09-30T12:00:00Z'
  /api/artistas/nombre/{nombre}/albumes:
    get:
      summary: Albums by artist name
      parameters:
        - name: nombre
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Albums found (may be empty)
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - properties:
                      data:
                        type: object
                        properties:
                          artista:
                            $ref: '#/components/schemas/Artist'
                          albumes:
                            type: array
                            items:
                              $ref: '#/components/schemas/Album'
        '400':
          description: Missing name
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '404':
          description: Artist not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '500':
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
  /api/albumes:
    get:
      summary: List albums
      responses:
        '200':
          description: Albums retrieved
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - properties:
                      data:
                        type: object
                        properties:
                          albumes:
                            type: array
                            items:
                              $ref: '#/components/schemas/Album'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
  /api/agregar_album:
    post:
      summary: Create album
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AlbumCreateInput'
      responses:
        '201':
          description: Album created
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - properties:
                      data:
                        type: object
                        properties:
                          album:
                            $ref: '#/components/schemas/Album'
        '400':
          description: Invalid payload
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '404':
          description: Artist not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '409':
          description: Duplicate album title
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '500':
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
  /api/albumes/{id}:
    patch:
      summary: Update album
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AlbumUpdateInput'
      responses:
        '200':
          description: Album updated
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - properties:
                      data:
                        type: object
                        properties:
                          album:
                            $ref: '#/components/schemas/Album'
        '400':
          description: Invalid payload
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '404':
          description: Album or referenced artist not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '409':
          description: Duplicate album title
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '500':
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'


    delete:
      summary: Delete album
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Album deleted
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - properties:
                      data:
                        type: object
                        properties:
                          album:
                            $ref: '#/components/schemas/Album'
              examples:
                albumDeleted:
                  summary: Eliminacion exitosa
                  value:
                    status: success
                    message: Album eliminado
                    data:
                      album:
                        id_album: 12
                        titulo_album: Eduardo
                        anio_album: 2021
                        id_artista: 26
                    timestamp: '2025-09-30T12:00:00Z'
        '400':
          description: Invalid id
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
              examples:
                invalidId:
                  summary: id no valido
                  value:
                    status: error
                    message: id debe ser un numero entero
                    data:
                      field: id
                    timestamp: '2025-09-30T12:00:00Z'
        '404':
          description: Album not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
              examples:
                notFound:
                  summary: Album inexistente
                  value:
                    status: error
                    message: Album no encontrado
                    data:
                      id_album: 99
                    timestamp: '2025-09-30T12:00:00Z'
        '409':
          description: Album has related records
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
              examples:
                hasDependencies:
                  summary: El album tiene dependencias
                  value:
                    status: error
                    message: No se puede eliminar el album porque tiene registros asociados
                    data:
                      id_album: 12
                    timestamp: '2025-09-30T12:00:00Z'
        '500':
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
              examples:
                serverError:
                  summary: Error inesperado
                  value:
                    status: error
                    message: Error al eliminar album
                    data:
                      error: Descripcion del error
                    timestamp: '2025-09-30T12:00:00Z'